# Accelerated `LCMAggregation` by increasing parallelism