I've vectorized the computations such that the total number of calls to `GRUCell` during a forward pass is reduced from $\mathcal{O}(|\mathcal{V}|)$ to $\mathcal{O}(\log |\mathcal{V}|)$. I used a benchmark script to compare the old version to the new one, and have introduced this benchmark into `test_lcm.py`. Here are the results, running the tests on an NVIDIA A100 GPU and AMD EPYC 7413 24-core CPU:

```
Old version:
+------------------+-----------+
| Name             | Forward   |
|------------------+-----------|
| N=32768, B=1024  | 3.5072s   |
| N=32768, B=2048  | 1.9874s   |
| N=32768, B=4096  | 1.8826s   |
| N=65536, B=1024  | 6.1835s   |
| N=65536, B=2048  | 3.6961s   |
| N=65536, B=4096  | 3.2131s   |
| N=131072, B=1024 | 11.3381s  |
| N=131072, B=2048 | 6.2320s   |
| N=131072, B=4096 | 5.2321s   |
| N=262144, B=1024 | 21.1921s  |
| N=262144, B=2048 | 11.6855s  |
| N=262144, B=4096 | 9.4793s   |
+------------------+-----------+

New version:
+------------------+-----------+
| Name             | Forward   |
|------------------+-----------|
| N=32768, B=1024  | 1.2499s   |
| N=32768, B=2048  | 1.3941s   |
| N=32768, B=4096  | 1.7324s   |
| N=65536, B=1024  | 1.9671s   |
| N=65536, B=2048  | 2.1052s   |
| N=65536, B=4096  | 2.5111s   |
| N=131072, B=1024 | 3.7243s   |
| N=131072, B=2048 | 3.7287s   |
| N=131072, B=4096 | 4.3217s   |
| N=262144, B=1024 | 7.0736s   |
| N=262144, B=2048 | 6.9741s   |
| N=262144, B=4096 | 7.1758s   |
+------------------+-----------+
```

We see up to $3\times$ speedup, and more consistent behavior when keeping `N` fixed and varying `B`.